requirements.txt
djangorestframework==3.9.2
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
flake8